requirements.txt
einops==0.4.1
facenet_pytorch==2.5.2
matplotlib
numpy
opencv_python==4.6.0.66
pandas==1.4.4
Pillow==9.2.0
scikit_learn
scipy
torch==2.1.2
torch_summary
torchvision
vit_pytorch
openpyxl
tqdm
opencv-python
facenet-pytorch
mediapipe
peakutils
onnx
imageio
python-dotenv
wandb
seaborn